refactor the monolith
