--- server-middleware/hodlers.json
@@ -1,1 +1,0 @@
-[{"discordName":"nft4cause#1140","publicKey":"8DyNeQYMWY6NLpPN7S1nTcDy2WXLnm5rzrtdWA2H2t6Y"}]
